# Conflicts:
#	main.cpp
#	src/Clases/Jugador/Jugador.h
#	src/Clases/Sistema.cpp
#	src/Clases/Videojuego/Videojuego.cpp

+++ src/Clases/Puntuacion/Puntuacion.cpp
@@ -2,23 +2,21 @@
 #define PUNTUACION_HEADER
 
 #include <iostream>
-#include "../../Clases/Videojuego/Videojuego.cpp"
-#include "../../Clases/Jugador/Jugador.cpp"
 
 using namespace std;
 
 class Puntuacion : public ICollectible
 {
 private:
-    int puntuacion;
+    double puntuacion;
     Videojuego *juego;
     Jugador *creador;
 
 public:
     Puntuacion(int, Videojuego *, Jugador *);
-    void setPuntuacion(int);
+    void setPuntuacion(double);
     void setCreador(Jugador *);
-    int getPuntuacion();
+    double getPuntuacion();
     Jugador *getCreador();
 };
 
@@ -29,7 +27,7 @@ Puntuacion::Puntuacion(int puntuacion, Videojuego *juego, Jugador *creador)
     this->creador = creador;
 }
 
-void Puntuacion::setPuntuacion(int puntuacion)
+void Puntuacion::setPuntuacion(double puntuacion)
 {
     this->puntuacion = puntuacion;
 }
@@ -39,7 +37,7 @@ void Puntuacion::setCreador(Jugador *creador)
     this->creador = creador;
 }
 
-int Puntuacion::getPuntuacion()
+double Puntuacion::getPuntuacion()
 {
     return this->puntuacion;
 }
